requirements.txt
# Django and REST Framework
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg[binary]==3.3.2
dj-database-url==2.1.0

# Authentication and JWT
PyJWT==2.8.0
python-jose[cryptography]==3.3.0

# Environment variables
python-decouple==3.8

# API Documentation
drf-spectacular==0.27.1

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Document Processing
PyPDF2==3.0.1
python-docx==0.8.11

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
